Module Yocaml_markdown

The recommended plugin for converting Markdown documents. It relies on the use of different packets to ensure different levels of conversion:

https://ocaml.org/p/cmarkit/latest

https://ocaml.org/p/hilite/latest for highlighting

val from_string_to_html : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> ?safe:bool -> string -> string

A shortcut to directly convert a Makrdown string into an HTML string.

Document

Allows you to compose directly with Cmarkit documents.

module Doc : sig ... end

Features related to direct document processing Cmarkit.

Pipeline

Concrete pipelines.

module Pipeline : sig ... end

Task to deal with Pipelines.