Module Yocaml_markdown.Doc

Features related to direct document processing Cmarkit.

val make : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> string -> Cmarkit.Doc.t

Produce a complete document. By default, highlight use the function syntax_highlighting.

val from_string : ?strict:bool -> ?heading_auto_ids:bool -> string -> Cmarkit.Doc.t

Converts a string into a Cmarkit document. If you want more parameters, you can still use the Cmarkit function.

val default_grammars_set : TmLanguage.t

Returns a set of default supported grammars set.

val table_of_contents : ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> Cmarkit.Doc.t -> string option

Compute the table of contents of a Markdown document (Where the first element of the couple is the ID of the section and the second is the VALUE).

val syntax_highlighting : ?skip_unknown_languages:bool -> ?tm:TmLanguage.t -> ?lookup_method:Hilite.tm_lookup_method -> unit -> Cmarkit.Doc.t -> Cmarkit.Doc.t

Highlight the given document using Hilite.

val no_highlighting : Cmarkit.Doc.t -> Cmarkit.Doc.t

To be used with make.

val to_html : ?safe:bool -> Cmarkit.Doc.t -> string

Converts a document to an HTML string.