Module Yocaml_markdown.Pipeline

Task to deal with Pipelines.

val make : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> ?safe:bool -> unit -> (string, string) Yocaml.Task.t

Convert a Markdown string to an HTML String.

val to_doc : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> unit -> (string, Cmarkit.Doc.t) Yocaml.Task.t

A Task for converting a string into a Cmarkit document.

val table_of_contents : ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> unit -> (Cmarkit.Doc.t, string option) Yocaml.Task.t

Compute the table of content of a given document.

val with_table_of_contents : ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> unit -> (Cmarkit.Doc.t, string option * Cmarkit.Doc.t) Yocaml.Task.t

Compute the table of contents of a given document and collapse the result.

val table_of_contents_metadata : ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> unit -> ('a * Cmarkit.Doc.t, ('a * string option) * Cmarkit.Doc.t) Yocaml.Task.t

Deal with table of contents at the metadata level.

val to_html : ?safe:bool -> unit -> (Cmarkit.Doc.t, string) Yocaml.Task.t

A Task to convert a document to HTML.

module With_metadata : sig ... end