Module Pipeline.With_metadata

val make : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> ?safe:bool -> unit -> ('a * string, 'a * string) Yocaml.Task.t

Convert a Markdown string to an HTML String taking metadata in account.

val make_with_table_of_contents : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> ?safe:bool -> unit -> ('a * string, ('a * string option) * string) Yocaml.Task.t

Same as make but with table of contents handling.

val to_doc : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> unit -> ('a * string, 'a * Cmarkit.Doc.t) Yocaml.Task.t

A Task for converting a string into a Cmarkit document.

val table_of_contents : ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> unit -> ('a * Cmarkit.Doc.t, ('a * string option) * Cmarkit.Doc.t) Yocaml.Task.t

val with_table_of_contents : ?strict:bool -> ?heading_auto_ids:bool -> ?highlight:(Cmarkit.Doc.t -> Cmarkit.Doc.t) -> ?traverse_table:((string * string) Yocaml.Markup.Toc.t -> string option) -> unit -> ('a * string, ('a * string option) * Cmarkit.Doc.t) Yocaml.Task.t

Convert a document and collapse the table of content into the metadata field.

val to_html : ?safe:bool -> unit -> ('a * Cmarkit.Doc.t, 'a * string) Yocaml.Task.t